+++ app/lib/common/CommonEvents.js
@@ -1,6 +1,7 @@
 export default {
-    DATACLASS_CREATION_REQUESTED: 'dataclass.create.requested',
-    STATE_CREATION_REQUESTED: 'state.create.requested',
-    NAME_CREATION_REQUESTED: 'name.create.requested',
-    ROLE_CREATION_REQUESTED: 'role.create.requested',
-}
+  DATACLASS_CREATION_REQUESTED: "dataclass.create.requested",
+  STATE_CREATION_REQUESTED: "state.create.requested",
+  NAME_CREATION_REQUESTED: "name.create.requested",
+  ROLE_CREATION_REQUESTED: "position.create.requested",
+  UNIT_CREATION_REQUESTED: "unit.create.requested",
+};
