+++ app/lib/fragmentmodeler/activityLabelHandling/index.js
@@ -1,0 +1,6 @@
+import ActivityLabelHandler from "./activityLabelHandler";
+
+export default {
+  __init__: ["activityLabelHandler"],
+  activityLabelHandler: ["type", ActivityLabelHandler],
+};
