+++ app/lib/fragmentmodeler/taskLabelHandling/taskLabelHandler.js
@@ -1,209 +1,538 @@
 import CommandInterceptor from "diagram-js/lib/command/CommandInterceptor";
 import CommonEvents from "../../common/CommonEvents";
 import getDropdown from "../../util/Dropdown";
-import {appendOverlayListeners} from "../../util/HtmlUtil";
-import {is} from "../../util/Util";
+import { appendOverlayListeners } from "../../util/HtmlUtil";
+import { without } from "min-dash";
+
+import { is } from "../../util/Util";
 
 export default class TaskLabelHandler extends CommandInterceptor {
-    constructor(eventBus, modeling, directEditing, overlays, fragmentModeler) {
-        super(eventBus);
-        this._eventBus = eventBus;
-        this._modeling = modeling;
-        this._dropdownContainer = document.createElement('div');
-        this._dropdownContainer.classList.add('dd-dropdown-multicontainer');
-        this._nameDropdown = getDropdown("Name");
-        this._dropdownContainer.appendChild(this._nameDropdown);
-        this._durationDropdown = getDropdown("Duration");
-        this._dropdownContainer.appendChild(this._durationDropdown);
-        this._roleDropdown = getDropdown("Role");
-        this._dropdownContainer.appendChild(this._roleDropdown);
-        this._NoPDropdown = getDropdown("Number Of People");
-        this._dropdownContainer.appendChild(this._NoPDropdown);
-        this._currentDropdownTarget = undefined;
-        this._overlayId = undefined;
-        this._overlays = overlays;
-        this._fragmentModeler = fragmentModeler;
+  constructor(eventBus, modeling, directEditing, overlays, resourceModeler) {
+    super(eventBus);
+    this._eventBus = eventBus;
+    this._modeling = modeling;
+    this._dropdownContainer = document.createElement("div");
+    this._dropdownContainer.classList.add("dd-dropdown-multicontainer");
 
-        eventBus.on('directEditing.activate', function (e) {
-            if (is(e.active.element, 'bpmn:Task')) {
-                directEditing.cancel();
+    this._rolesDropdown = getDropdown("Positions/Roles");
+    this._dropdownContainer.appendChild(this._rolesDropdown);
+    this._unitsDropdown = getDropdown("Organizational Units");
+    this._dropdownContainer.appendChild(this._unitsDropdown);
+
+    this._currentDropdownTarget = undefined;
+    this._overlayId = undefined;
+    this._overlays = overlays;
+    this._resourceModeler = resourceModeler;
+
+    eventBus.on("directEditing.activate", function (e) {
+      if (is(e.active.element, "bpmn:Participant")) {
+        directEditing.cancel();
+      }
+      if (is(e.active.element, "bpmn:Lane")) {
+        directEditing.cancel();
+      }
+    });
+
+    eventBus.on(["element.dblclick", "create.end", "autoPlace.end"], (e) => {
+      const element = e.element || e.shape || e.elements[0];
+      if (is(element, "bpmn:Participant")) {
+        const resource = element.businessObject;
+        this._dropdownContainer.currentElement = element;
+
+        const updateRolesSelection = () => {
+          // Clear selection for all entries
+          this._rolesDropdown
+            .getEntries()
+            .forEach((entry) => entry.setSelected(false));
+
+          // Set selected for the last entered role
+          const lastRole =
+            resource.roles && resource.roles.length > 0
+              ? resource.roles[resource.roles.length - 1]
+              : null;
+          if (lastRole) {
+            const selectedEntry = this._rolesDropdown
+              .getEntries()
+              .find((entry) => entry.option === lastRole);
+            if (selectedEntry) {
+              selectedEntry.setSelected(true);
             }
-        });
+          }
+        };
 
-        eventBus.on(['element.dblclick', 'create.end', 'autoPlace.end'], e => {
-            const element = e.element || e.shape || e.elements[0];
-            if (is(element, 'bpmn:Task')) {
-                const activity = element.businessObject;
-                this._dropdownContainer.currentElement = element;
+        const updateUnitsSelection = () => {
+          this._unitsDropdown
+            .getEntries()
+            .forEach((entry) =>
+              entry.setSelected(
+                resource.units?.find((unit) => unit === entry.option)
+              )
+            );
+        };
 
-                const updateRoleSelection = () => {
-                    this._roleDropdown.getEntries().forEach(entry => entry.setSelected(activity.role === entry.option));
-                }
+        const populateDropdown = (
+          dropdown,
+          items,
+          updateFunction,
+          updateSelectionFunction
+        ) => {
+          const newestItem = items.length > 0 ? items[items.length - 1] : null;
 
-                const populateNameDropdown = () => {
-                    this._nameDropdown.populate(
-                        [],
-                        () => {
-                        },
-                        element
-                    );
-                    this._nameDropdown.addCreateElementInput(event => this._dropdownContainer.confirm(), "text", activity.name);
-                }
-                const populateDurationDropdown = () => {
-                    this._durationDropdown.populate(
-                        [],
-                        () => {
-                        },
-                        element
-                    );
-                    this._durationDropdown.addCreateElementInput(event => this._dropdownContainer.confirm(), "number", activity.duration, "0");
-                }
-                const populateRoleDropdown = () => {
-                    this._roleDropdown.populate(
-                        this._fragmentModeler._roles || [],
-                        (role, element) => {
-                            this.updateRole(role, element);
-                            if (element.businessObject.role === undefined) {
-                                this._NoPDropdown.clearInput()
-                            }
-                            updateRoleSelection();
-                        },
-                        element
-                    );
-                    this._roleDropdown.addCreateElementInput(event => this._dropdownContainer.confirm());
-                    updateRoleSelection();
-                }
-                const populateNoPDropdown = () => {
-                    this._NoPDropdown.populate(
-                        [],
-                        () => {
-                        },
-                        element
-                    );
-                    this._NoPDropdown.addCreateElementInput(event => this._dropdownContainer.confirm(), "number", activity.NoP);
-                }
-                populateNameDropdown();
-                populateDurationDropdown();
-                populateRoleDropdown();
-                populateNoPDropdown();
+          dropdown.populate(
+            items,
+            (item, element) => {
+              updateFunction(item, element);
+              updateSelectionFunction();
+            },
+            element
+          );
 
-                this._dropdownContainer.confirm = (event) => {
-                    const newNameInput = this._nameDropdown.getInputValue().trim();
-                    const newDurationInput = this._durationDropdown.getInputValue().trim();
-                    const newRoleInput = this._roleDropdown.getInputValue().trim();
-                    const newNoPInput = this._NoPDropdown.getInputValue().trim();
-                    if (newNameInput !== '' && newNameInput !== activity.name) {
-                        this.updateName(newNameInput, element);
-                        populateNameDropdown();
-                    }
-                    if (newDurationInput !== activity.duration && newDurationInput >= 0) {
-                        this.updateDuration(newDurationInput, element);
-                        populateDurationDropdown();
-                    }
-                    if (newRoleInput !== '' && newRoleInput !== activity.role) {
-                        let newRole = this.createRole(newRoleInput);
-                        this.updateRole(newRole, element);
-                        populateRoleDropdown();
-                    }
-                    if (newNoPInput !== activity.NoP && newNoPInput > 0 && activity.role !== undefined) {
-                        this.updateNoP(newNoPInput, element);
-                        populateNoPDropdown();
-                    }
-                }
+          dropdown.getEntries().forEach((entry) => {
+            entry.addEventListener("click", (event) => {
+              const selectedItem = entry.option;
 
-                let shouldBlockNextClick = e.type === 'create.end';
-                this._dropdownContainer.handleClick = (event) => {
-                    if (shouldBlockNextClick) {
-                        shouldBlockNextClick = false;
-                        return true;
-                    } else if (!this._dropdownContainer.contains(event.target)) {
-                        return false;
-                    } else if (event.target.classList.contains('dd-dropdown-entry')) {
-                        this._roleDropdown.clearInput();
-                    } else if (event.target.tagName !== 'INPUT' || !event.target.value) {
-                        this._dropdownContainer.confirm();
-                    }
-                    return true;
-                }
+              // Clear selection for all entries in both dropdowns
+              this._rolesDropdown.getEntries().forEach((roleEntry) => {
+                roleEntry.setSelected(false);
+              });
 
-                this._dropdownContainer.close = () => {
-                    if (this._overlayId) {
-                        this._overlays.remove(this._overlayId);
-                        this._overlayId = undefined;
-                    }
-                    this._dropdownContainer.currentElement = undefined;
-                    this._currentDropdownTarget = undefined;
-                }
+              this._unitsDropdown.getEntries().forEach((unitEntry) => {
+                unitEntry.setSelected(false);
+              });
 
-                const closeOverlay = appendOverlayListeners(this._dropdownContainer);
-                eventBus.once('element.contextmenu', event => {
-                    if (this._currentDropdownTarget && ((event.element || event.shape).businessObject !== this._currentDropdownTarget)) {
-                        closeOverlay(event);
-                        event.preventDefault();
-                    }
+              // Update the function for the clicked dropdown
+              if (dropdown === this._rolesDropdown) {
+                this.updateRoles(selectedItem, element);
+                updateRolesSelection();
+                // Clear selection in Units dropdown
+                this._unitsDropdown.getEntries().forEach((unitEntry) => {
+                  unitEntry.setSelected(false);
+                });
+                // Set selected for the clicked entry in Units dropdown
+                entry.setSelected(true);
+              } else if (dropdown === this._unitsDropdown) {
+                this.updateUnits(selectedItem, element);
+                updateUnitsSelection();
+                // Clear selection in Roles dropdown
+                this._rolesDropdown.getEntries().forEach((roleEntry) => {
+                  roleEntry.setSelected(false);
                 });
 
-                // Show the menu(e)
-                this._overlayId = overlays.add(element.id, 'classSelection', {
-                    position: {
-                        bottom: 0,
-                        right: 0
-                    },
-                    scale: false,
-                    html: this._dropdownContainer
+                // Set selected for the clicked entry in Units dropdown
+                entry.setSelected(true);
+              }
+
+              this._dropdownContainer.confirm();
+            });
+          });
+
+          dropdown.addCreateElementInput((event) =>
+            this._dropdownContainer.confirm()
+          );
+
+          // Set selected state for the newest item
+          const newestEntry = dropdown
+            .getEntries()
+            .find((entry) => entry.option === newestItem);
+          if (newestEntry) {
+            newestEntry.setSelected(true);
+          }
+
+          updateSelectionFunction();
+        };
+
+        const populateRolesDropdown = () => {
+          const roles = this._resourceModeler._roles || [];
+          populateDropdown(
+            this._rolesDropdown,
+            roles,
+            (role, element) => {
+              this.updateRoles(role, element);
+            },
+            () => {
+              updateRolesSelection();
+            }
+          );
+        };
+
+        const populateUnitsDropdown = () => {
+          const units = this._resourceModeler._units || [];
+          populateDropdown(
+            this._unitsDropdown,
+            units,
+            (unit, element) => {
+              this.updateUnits(unit, element);
+            },
+            () => {
+              updateUnitsSelection();
+            }
+          );
+        };
+
+        populateRolesDropdown();
+        populateUnitsDropdown();
+
+        this._dropdownContainer.confirm = (event) => {
+          const newRoleInput = this._rolesDropdown.getInputValue().trim();
+
+          if (
+            newRoleInput !== "" &&
+            !this._resourceModeler._roles?.find(
+              (role) => role.name === newRoleInput
+            )
+          ) {
+            let newRole = this.createRole(newRoleInput);
+            this.updateRoles(newRole, element);
+            populateRolesDropdown();
+          }
+
+          const newUnitInput = this._unitsDropdown.getInputValue().trim();
+
+          if (
+            newUnitInput !== "" &&
+            !this._resourceModeler._units?.find(
+              (unit) => unit.name === newUnitInput
+            )
+          ) {
+            let newUnit = this.createUnit(newUnitInput);
+            this.updateUnits(newUnit, element);
+            populateUnitsDropdown();
+          }
+        };
+
+        // this._dropdownContainer.confirm = (event) => {
+        //   const newUnitInput = this._unitsDropdown.getInputValue().trim();
+
+        //   if (
+        //     newUnitInput !== "" &&
+        //     !this._resourceModeler._units?.find(
+        //       (unit) => unit.name === newUnitInput
+        //     )
+        //   ) {
+        //     let newUnit = this.createUnit(newUnitInput);
+        //     this.updateUnits(newUnit, element);
+        //     populateUnitsDropdown();
+        //   }
+        // };
+
+        let shouldBlockNextClick = e.type === "create.end";
+        this._dropdownContainer.handleClick = (event) => {
+          if (shouldBlockNextClick) {
+            shouldBlockNextClick = false;
+            return true;
+          } else if (!this._dropdownContainer.contains(event.target)) {
+            return false;
+          } else if (event.target.classList.contains("dd-dropdown-entry")) {
+            this._rolesDropdown.clearInput();
+          } else if (event.target.tagName !== "INPUT" || !event.target.value) {
+            this._dropdownContainer.confirm();
+          }
+          return true;
+        };
+
+        this._dropdownContainer.close = () => {
+          if (this._overlayId) {
+            this._overlays.remove(this._overlayId);
+            this._overlayId = undefined;
+          }
+          this._dropdownContainer.currentElement = undefined;
+          this._currentDropdownTarget = undefined;
+        };
+
+        const closeOverlay = appendOverlayListeners(this._dropdownContainer);
+        eventBus.once("element.contextmenu", (event) => {
+          if (
+            this._currentDropdownTarget &&
+            (event.element || event.shape).businessObject !==
+              this._currentDropdownTarget
+          ) {
+            closeOverlay(event);
+            event.preventDefault();
+          }
+        });
+
+        // Show the menu(e)
+        this._overlayId = overlays.add(element.id, "classSelection", {
+          position: {
+            bottom: 0,
+            right: 0,
+          },
+          scale: false,
+          html: this._dropdownContainer,
+        });
+
+        this._currentDropdownTarget = element.businessObject;
+      }
+
+      if (is(element, "bpmn:Lane")) {
+        const resource = element.businessObject;
+        this._dropdownContainer.currentElement = element;
+
+        const updateRolesSelection = () => {
+          this._rolesDropdown
+            .getEntries()
+            .forEach((entry) =>
+              entry.setSelected(
+                resource.roles?.find((role) => role === entry.option)
+              )
+            );
+        };
+
+        const updateUnitsSelection = () => {
+          this._unitsDropdown
+            .getEntries()
+            .forEach((entry) =>
+              entry.setSelected(
+                resource.units?.find((unit) => unit === entry.option)
+              )
+            );
+        };
+
+        const populateDropdown = (
+          dropdown,
+          items,
+          updateFunction,
+          updateSelectionFunction
+        ) => {
+          const newestItem = items.length > 0 ? items[items.length - 1] : null;
+
+          dropdown.populate(
+            items,
+            (item, element) => {
+              updateFunction(item, element);
+              updateSelectionFunction();
+            },
+            element
+          );
+
+          dropdown.getEntries().forEach((entry) => {
+            entry.addEventListener("click", (event) => {
+              const selectedItem = entry.option;
+
+              // Clear selection for all entries in both dropdowns
+              this._rolesDropdown.getEntries().forEach((roleEntry) => {
+                roleEntry.setSelected(false);
+              });
+
+              this._unitsDropdown.getEntries().forEach((unitEntry) => {
+                unitEntry.setSelected(false);
+              });
+
+              // Update the function for the clicked dropdown
+              if (dropdown === this._rolesDropdown) {
+                this.updateRoles(selectedItem, element);
+                updateRolesSelection();
+                // Clear selection in Units dropdown
+                this._unitsDropdown.getEntries().forEach((unitEntry) => {
+                  unitEntry.setSelected(false);
+                });
+                // Set selected for the clicked entry in Units dropdown
+                entry.setSelected(true);
+              } else if (dropdown === this._unitsDropdown) {
+                this.updateUnits(selectedItem, element);
+                updateUnitsSelection();
+                // Clear selection in Roles dropdown
+                this._rolesDropdown.getEntries().forEach((roleEntry) => {
+                  roleEntry.setSelected(false);
                 });
 
-                this._currentDropdownTarget = element.businessObject;
+                // Set selected for the clicked entry in Units dropdown
+                entry.setSelected(true);
+              }
+
+              this._dropdownContainer.confirm();
+            });
+          });
+
+          dropdown.addCreateElementInput((event) =>
+            this._dropdownContainer.confirm()
+          );
+
+          // Set selected state for the newest item
+          const newestEntry = dropdown
+            .getEntries()
+            .find((entry) => entry.option === newestItem);
+          if (newestEntry) {
+            newestEntry.setSelected(true);
+          }
+
+          updateSelectionFunction();
+        };
+
+        const populateRolesDropdown = () => {
+          const roles = this._resourceModeler._roles || [];
+          populateDropdown(
+            this._rolesDropdown,
+            roles,
+            (role, element) => {
+              this.updateRoles(role, element);
+            },
+            () => {
+              updateRolesSelection();
             }
-        });
-    }
+          );
+        };
 
-    updateName(newName, element) {
-        element.businessObject.name = newName;
-        this._eventBus.fire('element.changed', {
-            element
-        });
-    }
+        const populateUnitsDropdown = () => {
+          const units = this._resourceModeler._units || [];
+          populateDropdown(
+            this._unitsDropdown,
+            units,
+            (unit, element) => {
+              this.updateUnits(unit, element);
+            },
+            () => {
+              updateUnitsSelection();
+            }
+          );
+        };
 
-    updateDuration(newTime, element) {
-        element.businessObject.duration = newTime;
-        this._eventBus.fire('element.changed', {
-            element
-        });
-    }
+        populateRolesDropdown();
+        populateUnitsDropdown();
 
-    createRole(name) {
-        return this._eventBus.fire(CommonEvents.ROLE_CREATION_REQUESTED, {
-            name
-        });
-    }
+        this._dropdownContainer.confirm = (event) => {
+          const newRoleInput = this._rolesDropdown.getInputValue().trim();
 
-    updateRole(newRole, element) {
-        const fmObject = element.businessObject;
-        if (fmObject.role === newRole) {
-            fmObject.role = undefined;
-            fmObject.NoP = undefined;
-        } else {
-            fmObject.role = newRole;
-        }
-        this._eventBus.fire('element.changed', {
-            element
-        });
-    }
+          if (
+            newRoleInput !== "" &&
+            !this._resourceModeler._roles?.find(
+              (role) => role.name === newRoleInput
+            )
+          ) {
+            let newRole = this.createRole(newRoleInput);
+            this.updateRoles(newRole, element);
+            populateRolesDropdown();
+          }
 
-    updateNoP(newNoP, element) {
-        element.businessObject.NoP = newNoP;
-        this._eventBus.fire('element.changed', {
-            element
-        });
-    }
+          const newUnitInput = this._unitsDropdown.getInputValue().trim();
 
+          if (
+            newUnitInput !== "" &&
+            !this._resourceModeler._units?.find(
+              (unit) => unit.name === newUnitInput
+            )
+          ) {
+            let newUnit = this.createUnit(newUnitInput);
+            this.updateUnits(newUnit, element);
+            populateUnitsDropdown();
+          }
+        };
+
+        // this._dropdownContainer.confirm = (event) => {
+        //   const newUnitInput = this._unitsDropdown.getInputValue().trim();
+
+        //   if (
+        //     newUnitInput !== "" &&
+        //     !this._resourceModeler._units?.find(
+        //       (unit) => unit.name === newUnitInput
+        //     )
+        //   ) {
+        //     let newUnit = this.createUnit(newUnitInput);
+        //     this.updateUnits(newUnit, element);
+        //     populateUnitsDropdown();
+        //   }
+        // };
+
+        let shouldBlockNextClick = e.type === "create.end";
+        this._dropdownContainer.handleClick = (event) => {
+          if (shouldBlockNextClick) {
+            shouldBlockNextClick = false;
+            return true;
+          } else if (!this._dropdownContainer.contains(event.target)) {
+            return false;
+          } else if (event.target.classList.contains("dd-dropdown-entry")) {
+            this._rolesDropdown.clearInput();
+          } else if (event.target.tagName !== "INPUT" || !event.target.value) {
+            this._dropdownContainer.confirm();
+          }
+          return true;
+        };
+
+        this._dropdownContainer.close = () => {
+          if (this._overlayId) {
+            this._overlays.remove(this._overlayId);
+            this._overlayId = undefined;
+          }
+          this._dropdownContainer.currentElement = undefined;
+          this._currentDropdownTarget = undefined;
+        };
+
+        const closeOverlay = appendOverlayListeners(this._dropdownContainer);
+        eventBus.once("element.contextmenu", (event) => {
+          if (
+            this._currentDropdownTarget &&
+            (event.element || event.shape).businessObject !==
+              this._currentDropdownTarget
+          ) {
+            closeOverlay(event);
+            event.preventDefault();
+          }
+        });
+
+        // Show the menu(e)
+        this._overlayId = overlays.add(element.id, "classSelection", {
+          position: {
+            bottom: 0,
+            right: 0,
+          },
+          scale: false,
+          html: this._dropdownContainer,
+        });
+
+        this._currentDropdownTarget = element.businessObject;
+      }
+    });
+  }
+
+  createRole(name) {
+    return this._eventBus.fire(CommonEvents.ROLE_CREATION_REQUESTED, {
+      name,
+    });
+  }
+
+  updateRoles(newRole, element) {
+    if (element.businessObject.roles?.find((role) => role === newRole)) {
+      element.businessObject.roles = without(
+        element.businessObject.roles,
+        newRole
+      );
+    } else if (element.businessObject.roles) {
+      element.businessObject.roles.push(newRole);
+    } else {
+      element.businessObject.roles = [newRole];
+    }
+    this._eventBus.fire("element.changed", {
+      element,
+    });
+  }
+
+  createUnit(name) {
+    return this._eventBus.fire(CommonEvents.UNIT_CREATION_REQUESTED, {
+      name,
+    });
+  }
+
+  updateUnits(newUnit, element) {
+    if (element.businessObject.units?.find((unit) => unit === newUnit)) {
+      element.businessObject.units = without(
+        element.businessObject.units,
+        newUnit
+      );
+    } else if (element.businessObject.units) {
+      element.businessObject.units.push(newUnit);
+    } else {
+      element.businessObject.units = [newUnit];
+    }
+    this._eventBus.fire("element.changed", {
+      element,
+    });
+  }
 }
 
 TaskLabelHandler.$inject = [
-    'eventBus',
-    'modeling',
-    'directEditing',
-    'overlays',
-    'fragmentModeler'
+  "eventBus",
+  "modeling",
+  "directEditing",
+  "overlays",
+  "fragmentModeler",
 ];
