+++ app/lib/resourcemodeler/features/auto-place/ODAutoPlaceUtil.js
@@ -1,12 +1,13 @@
-import { is } from "../../../common/util/ModelUtil";
+import {is} from '../../../common/util/ModelUtil';
 
-import { asTRBL, getMid } from "diagram-js/lib/layout/LayoutUtil";
+import {asTRBL, getMid,} from 'diagram-js/lib/layout/LayoutUtil';
 
 import {
-  findFreePosition,
-  generateGetNextPosition,
-  getConnectedDistance,
-} from "diagram-js/lib/features/auto-place/AutoPlaceUtil";
+    findFreePosition,
+    generateGetNextPosition,
+    getConnectedDistance
+} from 'diagram-js/lib/features/auto-place/AutoPlaceUtil';
+
 
 /**
  * Find the new position for the target element to
@@ -18,9 +19,10 @@ import {
  * @return {Point}
  */
 export function getNewShapePosition(source, element) {
-  if (is(element, "rem:Position") || is(element, "rem:OrganizationalUnit")) {
-    return getFlowNodePosition(source, element);
-  }
+
+    if (is(element, 'rem:Resource')) {
+        return getFlowNodePosition(source, element);
+    }
 }
 
 /**
@@ -28,45 +30,42 @@ export function getNewShapePosition(source, element) {
  * compute actual distance from previous nodes in flow.
  */
 export function getFlowNodePosition(source, element) {
-  var sourceTrbl = asTRBL(source);
-  var sourceMid = getMid(source);
 
-  var horizontalDistance = getConnectedDistance(source, {
-    filter: function (connection) {
-      return is(connection, "rem:Link");
-    },
-  });
+    var sourceTrbl = asTRBL(source);
+    var sourceMid = getMid(source);
 
-  var margin = 30,
-    minDistance = 80,
-    orientation = "left";
+    var horizontalDistance = getConnectedDistance(source, {
+        filter: function (connection) {
+            return is(connection, 'rem:Link');
+        }
+    });
 
-  var position = {
-    x: sourceTrbl.right + horizontalDistance + element.width / 2,
-    y: sourceMid.y + getVerticalDistance(orientation, minDistance),
-  };
+    var margin = 30,
+        minDistance = 80,
+        orientation = 'left';
 
-  var nextPositionDirection = {
-    y: {
-      margin: margin,
-      minDistance: minDistance,
-    },
-  };
+    var position = {
+        x: sourceTrbl.right + horizontalDistance + element.width / 2,
+        y: sourceMid.y + getVerticalDistance(orientation, minDistance)
+    };
 
-  return findFreePosition(
-    source,
-    element,
-    position,
-    generateGetNextPosition(nextPositionDirection)
-  );
+    var nextPositionDirection = {
+        y: {
+            margin: margin,
+            minDistance: minDistance
+        }
+    };
+
+    return findFreePosition(source, element, position, generateGetNextPosition(nextPositionDirection));
 }
 
+
 function getVerticalDistance(orientation, minDistance) {
-  if (orientation.indexOf("top") != -1) {
-    return -1 * minDistance;
-  } else if (orientation.indexOf("bottom") != -1) {
-    return minDistance;
-  } else {
-    return 0;
-  }
+    if (orientation.indexOf('top') != -1) {
+        return -1 * minDistance;
+    } else if (orientation.indexOf('bottom') != -1) {
+        return minDistance;
+    } else {
+        return 0;
+    }
 }
