+++ app/lib/resourcemodeler/features/label-editing/LabelEditingProvider.js
@@ -1,129 +1,123 @@
-import { assign } from "min-dash";
+import {assign} from 'min-dash';
 
-import { getLabel } from "./LabelUtil";
+import {getLabel} from './LabelUtil';
 
-import { isAny } from "../../../common/features/modeling/ModelingUtil";
+import {isAny} from '../../../common/features/modeling/ModelingUtil';
 
-import {
-  getExternalLabelMid,
-  hasExternalLabel,
-  isLabel,
-  isLabelExternal,
-} from "../../util/LabelUtil";
+import {getExternalLabelMid, hasExternalLabel, isLabel, isLabelExternal} from '../../util/LabelUtil';
+
 
 export default function LabelEditingProvider(
-  eventBus,
-  odFactory,
-  canvas,
-  directEditing,
-  modeling,
-  resizeHandles,
-  textRenderer
-) {
-  this._odFactory = odFactory;
-  this._canvas = canvas;
-  this._modeling = modeling;
-  this._textRenderer = textRenderer;
+    eventBus, odFactory, canvas, directEditing,
+    modeling, resizeHandles, textRenderer) {
 
-  directEditing.registerProvider(this);
+    this._odFactory = odFactory;
+    this._canvas = canvas;
+    this._modeling = modeling;
+    this._textRenderer = textRenderer;
 
-  function decideIfTitelOrAttributesClicked(event) {
-    var zoom = canvas.zoom();
-    var titel_attribute_divider_y_coordinate =
-      (event.element.y + 30 - canvas._cachedViewbox.y) * zoom;
-    var click_y_coordinate = event.originalEvent.offsetY;
-    if (click_y_coordinate >= titel_attribute_divider_y_coordinate) {
-      event.element.businessObject.labelAttribute = "state";
-    } else {
-      event.element.businessObject.labelAttribute = "name";
-    }
-  }
+    directEditing.registerProvider(this);
 
-  // we do not need this reaction on double click anymore - it throws errors and infinite loops
-  // probably because our dropdown handles everything now
-
-  // listen to dblclick on non-root elements
-  // eventBus.on('element.dblclick', function(event) {
-  //   decideIfTitelOrAttributesClicked(event);
-  //   activateDirectEdit(event.element, true);
-  // });
-
-  // complete on followup canvas operation
-  eventBus.on(
-    [
-      "autoPlace.start",
-      "canvas.viewbox.changing",
-      "drag.init",
-      "element.mousedown",
-      "popupMenu.open",
-    ],
-    function (event) {
-      if (directEditing.isActive()) {
-        directEditing.complete();
-      }
-    }
-  );
-
-  // cancel on command stack changes
-  eventBus.on(["commandStack.changed"], function (e) {
-    if (directEditing.isActive()) {
-      directEditing.cancel();
-    }
-  });
-
-  eventBus.on("directEditing.activate", function (event) {
-    resizeHandles.removeResizers();
-  });
-
-  eventBus.on("create.end", 500, function (event) {
-    var context = event.context,
-      element = context.shape,
-      canExecute = event.context.canExecute,
-      isTouch = event.isTouch;
-
-    // TODO(nikku): we need to find a way to support the
-    // direct editing on mobile devices; right now this will
-    // break for desworkflowediting on mobile devices
-    // as it breaks the user interaction workflow
-
-    // TODO(nre): we should temporarily focus the edited element
-    // here and release the focused viewport after the direct edit
-    // operation is finished
-    if (isTouch) {
-      return;
+    function decideIfTitelOrAttributesClicked(event) {
+        var zoom = canvas.zoom();
+        var titel_attribute_divider_y_coordinate = (event.element.y + 30 - canvas._cachedViewbox.y) * zoom;
+        var click_y_coordinate = event.originalEvent.offsetY;
+        if (click_y_coordinate >= titel_attribute_divider_y_coordinate) {
+            event.element.businessObject.labelAttribute = 'state';
+        } else {
+            event.element.businessObject.labelAttribute = 'name';
+        }
     }
 
-    if (!canExecute) {
-      return;
+    // we do not need this reaction on double click anymore - it throws errors and infinite loops
+    // probably because our dropdown handles everything now
+
+    // listen to dblclick on non-root elements
+    // eventBus.on('element.dblclick', function(event) {
+    //   decideIfTitelOrAttributesClicked(event);
+    //   activateDirectEdit(event.element, true);
+    // });
+
+    // complete on followup canvas operation
+    eventBus.on([
+        'autoPlace.start',
+        'canvas.viewbox.changing',
+        'drag.init',
+        'element.mousedown',
+        'popupMenu.open'
+    ], function (event) {
+
+        if (directEditing.isActive()) {
+            directEditing.complete();
+        }
+    });
+
+    // cancel on command stack changes
+    eventBus.on(['commandStack.changed'], function (e) {
+        if (directEditing.isActive()) {
+            directEditing.cancel();
+        }
+    });
+
+
+    eventBus.on('directEditing.activate', function (event) {
+        resizeHandles.removeResizers();
+    });
+
+    eventBus.on('create.end', 500, function (event) {
+
+        var context = event.context,
+            element = context.shape,
+            canExecute = event.context.canExecute,
+            isTouch = event.isTouch;
+
+        // TODO(nikku): we need to find a way to support the
+        // direct editing on mobile devices; right now this will
+        // break for desworkflowediting on mobile devices
+        // as it breaks the user interaction workflow
+
+        // TODO(nre): we should temporarily focus the edited element
+        // here and release the focused viewport after the direct edit
+        // operation is finished
+        if (isTouch) {
+            return;
+        }
+
+        if (!canExecute) {
+            return;
+        }
+
+        if (context.hints && context.hints.createElementsBehavior === false) {
+            return;
+        }
+
+        activateDirectEdit(element, false);
+    });
+
+    eventBus.on('autoPlace.end', 500, function (event) {
+        activateDirectEdit(event.shape, false);
+    });
+
+
+    function activateDirectEdit(element, force) {
+        if (force ||
+            isAny(element, ['rem:Resource'])) {
+            directEditing.activate(element);
+        }
     }
 
-    if (context.hints && context.hints.createElementsBehavior === false) {
-      return;
-    }
-
-    activateDirectEdit(element, false);
-  });
-
-  eventBus.on("autoPlace.end", 500, function (event) {
-    activateDirectEdit(event.shape, false);
-  });
-
-  function activateDirectEdit(element, force) {
-    if (force || isAny(element, ["rem:Position", "rem:OrganizationalUnit"])) {
-      directEditing.activate(element);
-    }
-  }
 }
 
 LabelEditingProvider.$inject = [
-  "eventBus",
-  "odFactory",
-  "canvas",
-  "directEditing",
-  "modeling",
-  "resizeHandles",
-  "textRenderer",
+    'eventBus',
+    'odFactory',
+    'canvas',
+    'directEditing',
+    'modeling',
+    'resizeHandles',
+    'textRenderer'
 ];
+
 
 /**
  * Activate direct editing for objects and text annotations.
@@ -133,37 +127,39 @@ LabelEditingProvider.$inject = [
  * @return {Object} an object with properties bounds (position and size), text and options
  */
 LabelEditingProvider.prototype.activate = function (element) {
-  // text
-  var text = getLabel(element);
 
-  if (text === undefined) {
-    return;
-  }
+    // text
+    var text = getLabel(element);
 
-  var context = {
-    text: text,
-  };
+    if (text === undefined) {
+        return;
+    }
 
-  // bounds
-  var bounds = this.getEditingBBox(element);
+    var context = {
+        text: text
+    };
 
-  assign(context, bounds);
+    // bounds
+    var bounds = this.getEditingBBox(element);
 
-  var options = {};
+    assign(context, bounds);
 
-  // external labels
-  if (isLabelExternal(element)) {
-    assign(options, {
-      autoResize: true,
+    var options = {};
+
+    // external labels
+    if (isLabelExternal(element)) {
+        assign(options, {
+            autoResize: true
+        });
+    }
+
+    assign(context, {
+        options: options
     });
-  }
 
-  assign(context, {
-    options: options,
-  });
-
-  return context;
+    return context;
 };
+
 
 /**
  * Get the editing bounding box based on the element's size and position
@@ -174,132 +170,137 @@ LabelEditingProvider.prototype.activate = function (element) {
  *                  and size (fixed or minimum and/or maximum)
  */
 LabelEditingProvider.prototype.getEditingBBox = function (element) {
-  var canvas = this._canvas;
+    var canvas = this._canvas;
 
-  var target = element.label || element;
+    var target = element.label || element;
 
-  var bbox = canvas.getAbsoluteBBox(target);
+    var bbox = canvas.getAbsoluteBBox(target);
 
-  var mid = {
-    x: bbox.x + bbox.width / 2,
-    y: bbox.y + bbox.height / 2,
-  };
+    var mid = {
+        x: bbox.x + bbox.width / 2,
+        y: bbox.y + bbox.height / 2
+    };
 
-  // default position
-  var bounds = { x: bbox.x, y: bbox.y };
+    // default position
+    var bounds = {x: bbox.x, y: bbox.y};
 
-  var zoom = canvas.zoom();
+    var zoom = canvas.zoom();
 
-  var defaultStyle = this._textRenderer.getDefaultStyle(),
-    externalStyle = this._textRenderer.getExternalStyle();
+    var defaultStyle = this._textRenderer.getDefaultStyle(),
+        externalStyle = this._textRenderer.getExternalStyle();
 
-  // take zoom into account
-  var externalFontSize = externalStyle.fontSize * zoom,
-    externalLineHeight = externalStyle.lineHeight,
-    defaultFontSize = defaultStyle.fontSize * zoom,
-    defaultLineHeight = defaultStyle.lineHeight;
+    // take zoom into account
+    var externalFontSize = externalStyle.fontSize * zoom,
+        externalLineHeight = externalStyle.lineHeight,
+        defaultFontSize = defaultStyle.fontSize * zoom,
+        defaultLineHeight = defaultStyle.lineHeight;
 
-  var style = {
-    fontFamily: this._textRenderer.getDefaultStyle().fontFamily,
-    fontWeight: this._textRenderer.getDefaultStyle().fontWeight,
-  };
+    var style = {
+        fontFamily: this._textRenderer.getDefaultStyle().fontFamily,
+        fontWeight: this._textRenderer.getDefaultStyle().fontWeight
+    };
 
-  if (isAny(element, ["rem:Position", "rem:OrganizationalUnit"])) {
-    assign(bounds, {
-      width: bbox.width,
-      height: bbox.height,
-    });
 
-    assign(style, {
-      fontSize: defaultFontSize + "px",
-      lineHeight: defaultLineHeight,
-      paddingTop: 7 * zoom + "px",
-      paddingBottom: 7 * zoom + "px",
-      paddingLeft: 5 * zoom + "px",
-      paddingRight: 5 * zoom + "px",
-    });
+    if (isAny(element, ['rem:Resource'])) {
 
-    if (isAny(element, ["rem:Position", "rem:OrganizationalUnit"])) {
-      // Editing attributes should be different.
-      if (element.businessObject.labelAttribute === "state") {
         assign(bounds, {
-          y: bbox.y + 30 * zoom,
-          height: bbox.height - 30 * zoom,
+            width: bbox.width,
+            height: bbox.height
         });
-      } else {
-        assign(bounds, {
-          height: 30 * zoom,
+
+        assign(style, {
+            fontSize: defaultFontSize + 'px',
+            lineHeight: defaultLineHeight,
+            paddingTop: (7 * zoom) + 'px',
+            paddingBottom: (7 * zoom) + 'px',
+            paddingLeft: (5 * zoom) + 'px',
+            paddingRight: (5 * zoom) + 'px'
         });
-      }
+
+        if (isAny(element, ['rem:Resource'])) {
+
+            // Editing attributes should be different.
+            if (element.businessObject.labelAttribute === 'state') {
+                assign(bounds, {
+                    y: bbox.y + (30 * zoom),
+                    height: bbox.height - (30 * zoom)
+                });
+            } else {
+                assign(bounds, {
+                    height: (30 * zoom)
+                });
+            }
+        }
     }
-  }
 
-  var width = 90 * zoom,
-    paddingTop = 7 * zoom,
-    paddingBottom = 4 * zoom;
+    var width = 90 * zoom,
+        paddingTop = 7 * zoom,
+        paddingBottom = 4 * zoom;
 
-  // external labels for events, data elements, gateways, groups and connections
-  if (target.labelTarget) {
-    assign(bounds, {
-      width: width,
-      height: bbox.height + paddingTop + paddingBottom,
-      x: mid.x - width / 2,
-      y: bbox.y - paddingTop,
-    });
+    // external labels for events, data elements, gateways, groups and connections
+    if (target.labelTarget) {
+        assign(bounds, {
+            width: width,
+            height: bbox.height + paddingTop + paddingBottom,
+            x: mid.x - width / 2,
+            y: bbox.y - paddingTop
+        });
 
-    assign(style, {
-      fontSize: externalFontSize + "px",
-      lineHeight: externalLineHeight,
-      paddingTop: paddingTop + "px",
-      paddingBottom: paddingBottom + "px",
-    });
-  }
+        assign(style, {
+            fontSize: externalFontSize + 'px',
+            lineHeight: externalLineHeight,
+            paddingTop: paddingTop + 'px',
+            paddingBottom: paddingBottom + 'px'
+        });
+    }
 
-  // external label not yet created
-  if (
-    isLabelExternal(target) &&
-    !hasExternalLabel(target) &&
-    !isLabel(target)
-  ) {
-    var externalLabelMid = getExternalLabelMid(element);
+    // external label not yet created
+    if (isLabelExternal(target)
+        && !hasExternalLabel(target)
+        && !isLabel(target)) {
 
-    var absoluteBBox = canvas.getAbsoluteBBox({
-      x: externalLabelMid.x,
-      y: externalLabelMid.y,
-      width: 0,
-      height: 0,
-    });
+        var externalLabelMid = getExternalLabelMid(element);
 
-    var height = externalFontSize + paddingTop + paddingBottom;
+        var absoluteBBox = canvas.getAbsoluteBBox({
+            x: externalLabelMid.x,
+            y: externalLabelMid.y,
+            width: 0,
+            height: 0
+        });
 
-    assign(bounds, {
-      width: width,
-      height: height,
-      x: absoluteBBox.x - width / 2,
-      y: absoluteBBox.y - height / 2,
-    });
+        var height = externalFontSize + paddingTop + paddingBottom;
 
-    assign(style, {
-      fontSize: externalFontSize + "px",
-      lineHeight: externalLineHeight,
-      paddingTop: paddingTop + "px",
-      paddingBottom: paddingBottom + "px",
-    });
-  }
+        assign(bounds, {
+            width: width,
+            height: height,
+            x: absoluteBBox.x - width / 2,
+            y: absoluteBBox.y - height / 2
+        });
 
-  return { bounds: bounds, style: style };
+        assign(style, {
+            fontSize: externalFontSize + 'px',
+            lineHeight: externalLineHeight,
+            paddingTop: paddingTop + 'px',
+            paddingBottom: paddingBottom + 'px'
+        });
+    }
+
+    return {bounds: bounds, style: style};
 };
+
 
 LabelEditingProvider.prototype.update = function (element, newLabel) {
-  if (isEmptyText(newLabel)) {
-    newLabel = null;
-  }
 
-  this._modeling.updateLabel(element, newLabel);
+    if (isEmptyText(newLabel)) {
+        newLabel = null;
+    }
+
+    this._modeling.updateLabel(element, newLabel);
 };
+
 
 // helpers //////////////////////
 
 function isEmptyText(label) {
-  return !label || !label.trim();
+    return !label || !label.trim();
 }
