+++ app/lib/resourcemodeler/features/label-editing/LabelUtil.js
@@ -1,30 +1,31 @@
-import { isAny } from "../../../common/features/modeling/ModelingUtil";
+import {isAny} from '../../../common/features/modeling/ModelingUtil';
 
 function getLabelAttr(semantic) {
-  if (semantic.labelAttribute) {
-    return semantic.labelAttribute;
-  }
-  if (isAny(semantic, ["rem:Link", "rem:Position", "rem:OrganizationalUnit"])) {
-    return "name";
-  }
+    if (semantic.labelAttribute) {
+        return semantic.labelAttribute;
+    }
+    if (isAny(semantic, ['rem:Link', 'rem:Resource'])) {
+        return 'name';
+    }
 }
 
 export function getLabel(element) {
-  var semantic = element.businessObject;
-  var attr = getLabelAttr(semantic);
+    var semantic = element.businessObject;
+    var attr = getLabelAttr(semantic);
 
-  if (attr) {
-    return semantic[attr] || "";
-  }
+    if (attr) {
+        return semantic[attr] || '';
+    }
 }
 
+
 export function setLabel(element, text) {
-  var semantic = element.businessObject,
-    attr = getLabelAttr(semantic);
+    var semantic = element.businessObject,
+        attr = getLabelAttr(semantic);
 
-  if (attr) {
-    semantic[attr] = text;
-  }
+    if (attr) {
+        semantic[attr] = text;
+    }
 
-  return element;
+    return element;
 }
