+++ app/lib/resourcemodeler/features/modeling/ODFactory.js
@@ -1,101 +1,96 @@
-import { assign, map, pick } from "min-dash";
+import {assign, map, pick,} from 'min-dash';
 
-import { isAny } from "../../../common/features/modeling/ModelingUtil";
+import {isAny} from '../../../common/features/modeling/ModelingUtil';
 
-import { is } from "../../../common/util/ModelUtil";
+import {is} from '../../../common/util/ModelUtil';
+
 
 export default function ODFactory(moddle) {
-  this._model = moddle;
+    this._model = moddle;
 }
 
-ODFactory.$inject = ["moddle"];
+ODFactory.$inject = ['moddle'];
+
 
 ODFactory.prototype._needsId = function (element) {
-  return isAny(element, [
-    "rem:BoardElement",
-    "odDi:OdRootBoard",
-    "odDi:OdPlane",
-  ]);
+    return isAny(element, [
+        'rem:BoardElement', 'odDi:OdRootBoard', 'odDi:OdPlane'
+    ]);
 };
 
 ODFactory.prototype._ensureId = function (element) {
-  // generate semantic ids for elements
-  // rem:Object -> Object_ID
-  var prefix;
 
-  if (is(element, "rem:Position") || is(element, "rem:OrganizationalUnit")) {
-    prefix = "Object";
-  } else {
-    prefix = (element.$type || "").replace(/^[^:]*:/g, "");
-  }
+    // generate semantic ids for elements
+    // rem:Object -> Object_ID
+    var prefix;
 
-  prefix += "_";
+    if (is(element, 'rem:Resource')) {
+        prefix = 'Object';
+    } else {
+        prefix = (element.$type || '').replace(/^[^:]*:/g, '');
+    }
 
-  if (!element.id && this._needsId(element)) {
-    element.id = this._model.ids.nextPrefixed(prefix, element);
-  }
+    prefix += '_';
+
+    if (!element.id && this._needsId(element)) {
+        element.id = this._model.ids.nextPrefixed(prefix, element);
+    }
 };
+
 
 ODFactory.prototype.create = function (type, attrs) {
-  var element = this._model.create(type, attrs || {});
-  if (type === "rem:Position" || type === "rem:OrganizationalUnit") {
-    element.state = null;
-  }
+    var element = this._model.create(type, attrs || {});
+    if (type === 'rem:Resource') {
+        element.state = null;
+    }
 
-  this._ensureId(element);
+    this._ensureId(element);
 
-  return element;
+    return element;
 };
+
 
 ODFactory.prototype.createDiLabel = function () {
-  return this.create("odDi:OdLabel", {
-    bounds: this.createDiBounds(),
-  });
+    return this.create('odDi:OdLabel', {
+        bounds: this.createDiBounds()
+    });
 };
+
 
 ODFactory.prototype.createDiShape = function (semantic, bounds, attrs) {
-  return this.create(
-    "odDi:OdShape",
-    assign(
-      {
+
+    return this.create('odDi:OdShape', assign({
         boardElement: semantic,
-        bounds: this.createDiBounds(bounds),
-      },
-      attrs
-    )
-  );
+        bounds: this.createDiBounds(bounds)
+    }, attrs));
 };
 
+
 ODFactory.prototype.createDiBounds = function (bounds) {
-  return this.create("dc:Bounds", bounds);
+    return this.create('dc:Bounds', bounds);
 };
 
 ODFactory.prototype.createDiEdge = function (semantic, waypoints, attrs) {
-  return this.create(
-    "odDi:Link",
-    assign(
-      {
-        boardElement: semantic,
-      },
-      attrs
-    )
-  );
+    return this.create('odDi:Link', assign({
+        boardElement: semantic
+    }, attrs));
 };
 
+
 ODFactory.prototype.createDiPlane = function (semantic) {
-  return this.create("odDi:OdPlane", {
-    boardElement: semantic,
-  });
+    return this.create('odDi:OdPlane', {
+        boardElement: semantic
+    });
 };
 
 ODFactory.prototype.createDiWaypoints = function (waypoints) {
-  var self = this;
+    var self = this;
 
-  return map(waypoints, function (pos) {
-    return self.createDiWaypoint(pos);
-  });
+    return map(waypoints, function (pos) {
+        return self.createDiWaypoint(pos);
+    });
 };
 
 ODFactory.prototype.createDiWaypoint = function (point) {
-  return this.create("dc:Point", pick(point, ["x", "y"]));
+    return this.create('dc:Point', pick(point, ['x', 'y']));
 };
