+++ app/lib/resourcemodeler/features/palette/PaletteProvider.js
@@ -1,127 +1,116 @@
-import { assign } from "min-dash";
+import {assign} from 'min-dash';
 
 /**
  * A palette provider for od elements.
  */
 export default function PaletteProvider(
-  palette,
-  create,
-  elementFactory,
-  spaceTool,
-  lassoTool,
-  handTool,
-  globalConnect,
-  translate
-) {
-  this._create = create;
-  this._elementFactory = elementFactory;
-  this._spaceTool = spaceTool;
-  this._lassoTool = lassoTool;
-  this._handTool = handTool;
-  this._globalConnect = globalConnect;
-  this._translate = translate;
+    palette, create, elementFactory,
+    spaceTool, lassoTool, handTool, globalConnect, translate) {
 
-  palette.registerProvider(this);
+    this._create = create;
+    this._elementFactory = elementFactory;
+    this._spaceTool = spaceTool;
+    this._lassoTool = lassoTool;
+    this._handTool = handTool;
+    this._globalConnect = globalConnect;
+    this._translate = translate;
+
+    palette.registerProvider(this);
 }
 
 PaletteProvider.$inject = [
-  "palette",
-  "create",
-  "elementFactory",
-  "spaceTool",
-  "lassoTool",
-  "handTool",
-  "globalConnect",
-  "translate",
+    'palette',
+    'create',
+    'elementFactory',
+    'spaceTool',
+    'lassoTool',
+    'handTool',
+    'globalConnect',
+    'translate'
 ];
 
-PaletteProvider.prototype.getPaletteEntries = function (element) {
-  var actions = {},
-    create = this._create,
-    elementFactory = this._elementFactory,
-    spaceTool = this._spaceTool,
-    lassoTool = this._lassoTool,
-    handTool = this._handTool,
-    globalConnect = this._globalConnect,
-    translate = this._translate;
 
-  function createAction(type, group, className, title, options) {
-    function createListener(event) {
-      var shape = elementFactory.createShape(assign({ type: type }, options));
-      create.start(event, shape);
+PaletteProvider.prototype.getPaletteEntries = function (element) {
+
+    var actions = {},
+        create = this._create,
+        elementFactory = this._elementFactory,
+        spaceTool = this._spaceTool,
+        lassoTool = this._lassoTool,
+        handTool = this._handTool,
+        globalConnect = this._globalConnect,
+        translate = this._translate;
+
+    function createAction(type, group, className, title, options) {
+
+        function createListener(event) {
+            var shape = elementFactory.createShape(assign({type: type}, options));
+            create.start(event, shape);
+        }
+
+        var shortType = type.replace(/^rem:/, '');
+
+        return {
+            group: group,
+            className: className,
+            title: title || translate('Create {type}', {type: shortType}),
+            action: {
+                dragstart: createListener,
+                click: createListener
+            }
+        };
     }
 
-    var shortType = type.replace(/^rem:/, "");
-
-    return {
-      group: group,
-      className: className,
-      title: title || translate("Create {type}", { type: shortType }),
-      action: {
-        dragstart: createListener,
-        click: createListener,
-      },
-    };
-  }
-
-  assign(actions, {
-    "hand-tool": {
-      group: "tools",
-      className: "bpmn-icon-hand-tool",
-      title: translate("Activate the hand tool"),
-      action: {
-        click: function (event) {
-          handTool.activateHand(event);
+    assign(actions, {
+        'hand-tool': {
+            group: 'tools',
+            className: 'bpmn-icon-hand-tool',
+            title: translate('Activate the hand tool'),
+            action: {
+                click: function (event) {
+                    handTool.activateHand(event);
+                }
+            }
         },
-      },
-    },
-    "lasso-tool": {
-      group: "tools",
-      className: "bpmn-icon-lasso-tool",
-      title: translate("Activate the lasso tool"),
-      action: {
-        click: function (event) {
-          lassoTool.activateSelection(event);
+        'lasso-tool': {
+            group: 'tools',
+            className: 'bpmn-icon-lasso-tool',
+            title: translate('Activate the lasso tool'),
+            action: {
+                click: function (event) {
+                    lassoTool.activateSelection(event);
+                }
+            }
         },
-      },
-    },
-    "space-tool": {
-      group: "tools",
-      className: "bpmn-icon-space-tool",
-      title: translate("Activate the create/remove space tool"),
-      action: {
-        click: function (event) {
-          spaceTool.activateSelection(event);
+        'space-tool': {
+            group: 'tools',
+            className: 'bpmn-icon-space-tool',
+            title: translate('Activate the create/remove space tool'),
+            action: {
+                click: function (event) {
+                    spaceTool.activateSelection(event);
+                }
+            }
         },
-      },
-    },
-    "tool-separator": {
-      group: "tools",
-      separator: true,
-    },
-    "create-position-role-object": createAction(
-      "rem:Position",
-      "od-elements",
-      "bpmn-icon-task",
-      translate("Create Position/Role")
-    ),
-    "create-organizational-unit-object": createAction(
-      "rem:OrganizationalUnit",
-      "od-elements",
-      "bpmn-icon-transaction",
-      translate("Create Organizational Unit")
-    ),
-    // 'object-linker': {
-    //     group: 'od-elements',
-    //     className: 'bpmn-icon-connection',
-    //     title: translate('Link objects'),
-    //     action: {
-    //         click: function (event) {
-    //             globalConnect.start(event);
-    //         }
-    //     }
-    // }
-  });
+        'tool-separator': {
+            group: 'tools',
+            separator: true
+        },
+        'create-object': createAction(
+            'rem:Resource', 'od-elements', 'bpmn-icon-od-class',
+            translate('Create object')
+        ),
+        'object-linker': {
+            group: 'od-elements',
+            className: 'bpmn-icon-connection',
+            title: translate('Link objects'),
+            action: {
+                click: function (event) {
+                    globalConnect.start(event);
+                }
+            }
+        }
+    });
 
-  return actions;
+    return actions;
 };
