+++ app/lib/rolemodeler/features/auto-place/ODAutoPlaceUtil.js
@@ -1,13 +1,12 @@
-import {is} from '../../../common/util/ModelUtil';
+import { is } from "../../../common/util/ModelUtil";
 
-import {asTRBL, getMid,} from 'diagram-js/lib/layout/LayoutUtil';
+import { asTRBL, getMid } from "diagram-js/lib/layout/LayoutUtil";
 
 import {
-    findFreePosition,
-    generateGetNextPosition,
-    getConnectedDistance
-} from 'diagram-js/lib/features/auto-place/AutoPlaceUtil';
-
+  findFreePosition,
+  generateGetNextPosition,
+  getConnectedDistance,
+} from "diagram-js/lib/features/auto-place/AutoPlaceUtil";
 
 /**
  * Find the new position for the target element to
@@ -19,10 +18,9 @@ import {
  * @return {Point}
  */
 export function getNewShapePosition(source, element) {
-
-    if (is(element, 'rom:Role')) {
-        return getFlowNodePosition(source, element);
-    }
+  if (is(element, "rom:Position") || is(element, "rom:OrganizationalUnit")) {
+    return getFlowNodePosition(source, element);
+  }
 }
 
 /**
@@ -30,42 +28,45 @@ export function getNewShapePosition(source, element) {
  * compute actual distance from previous nodes in flow.
  */
 export function getFlowNodePosition(source, element) {
+  var sourceTrbl = asTRBL(source);
+  var sourceMid = getMid(source);
 
-    var sourceTrbl = asTRBL(source);
-    var sourceMid = getMid(source);
+  var horizontalDistance = getConnectedDistance(source, {
+    filter: function (connection) {
+      return is(connection, "rom:Inheritance");
+    },
+  });
 
-    var horizontalDistance = getConnectedDistance(source, {
-        filter: function (connection) {
-            return is(connection, 'rom:Inheritance');
-        }
-    });
+  var margin = 30,
+    minDistance = 80,
+    orientation = "left";
 
-    var margin = 30,
-        minDistance = 80,
-        orientation = 'left';
+  var position = {
+    x: sourceTrbl.right + horizontalDistance + element.width / 2,
+    y: sourceMid.y + getVerticalDistance(orientation, minDistance),
+  };
 
-    var position = {
-        x: sourceTrbl.right + horizontalDistance + element.width / 2,
-        y: sourceMid.y + getVerticalDistance(orientation, minDistance)
-    };
+  var nextPositionDirection = {
+    y: {
+      margin: margin,
+      minDistance: minDistance,
+    },
+  };
 
-    var nextPositionDirection = {
-        y: {
-            margin: margin,
-            minDistance: minDistance
-        }
-    };
-
-    return findFreePosition(source, element, position, generateGetNextPosition(nextPositionDirection));
+  return findFreePosition(
+    source,
+    element,
+    position,
+    generateGetNextPosition(nextPositionDirection)
+  );
 }
 
-
 function getVerticalDistance(orientation, minDistance) {
-    if (orientation.indexOf('top') != -1) {
-        return -1 * minDistance;
-    } else if (orientation.indexOf('bottom') != -1) {
-        return minDistance;
-    } else {
-        return 0;
-    }
+  if (orientation.indexOf("top") != -1) {
+    return -1 * minDistance;
+  } else if (orientation.indexOf("bottom") != -1) {
+    return minDistance;
+  } else {
+    return 0;
+  }
 }
