+++ app/lib/rolemodeler/features/label-editing/LabelEditingProvider.js
@@ -1,129 +1,131 @@
-import {assign} from 'min-dash';
+import { assign } from "min-dash";
 
-import {getLabel} from './LabelUtil';
+import { getLabel } from "./LabelUtil";
 
-import {isAny} from '../../../common/features/modeling/ModelingUtil';
+import { isAny } from "../../../common/features/modeling/ModelingUtil";
 
-import {getExternalLabelMid, hasExternalLabel, isLabel, isLabelExternal} from '../../util/LabelUtil';
-
+import {
+  getExternalLabelMid,
+  hasExternalLabel,
+  isLabel,
+  isLabelExternal,
+} from "../../util/LabelUtil";
 
 export default function LabelEditingProvider(
-    eventBus, odFactory, canvas, directEditing,
-    modeling, resizeHandles, textRenderer) {
+  eventBus,
+  odFactory,
+  canvas,
+  directEditing,
+  modeling,
+  resizeHandles,
+  textRenderer
+) {
+  this._odFactory = odFactory;
+  this._canvas = canvas;
+  this._modeling = modeling;
+  this._textRenderer = textRenderer;
+  this._eventBus = eventBus;
 
-    this._odFactory = odFactory;
-    this._canvas = canvas;
-    this._modeling = modeling;
-    this._textRenderer = textRenderer;
-    this._eventBus = eventBus;
+  directEditing.registerProvider(this);
 
-    directEditing.registerProvider(this);
+  // listen to dblclick on non-root elements
+  eventBus.on("element.dblclick", function (event) {
+    // event.element.businessObject.labelAttribute = 'name';
+    activateDirectEdit(event.element, true);
+  });
 
-    // listen to dblclick on non-root elements
-    eventBus.on('element.dblclick', function (event) {
-        // event.element.businessObject.labelAttribute = 'name';
-        activateDirectEdit(event.element, true);
-    });
+  // complete on followup canvas operation
+  eventBus.on(
+    [
+      "autoPlace.start",
+      "canvas.viewbox.changing",
+      "drag.init",
+      "element.mousedown",
+      "popupMenu.open",
+    ],
+    function (event) {
+      if (directEditing.isActive()) {
+        directEditing.complete();
+      }
+    }
+  );
 
-    // complete on followup canvas operation
-    eventBus.on([
-        'autoPlace.start',
-        'canvas.viewbox.changing',
-        'drag.init',
-        'element.mousedown',
-        'popupMenu.open'
-    ], function (event) {
+  // cancel on command stack changes
+  eventBus.on(["commandStack.changed"], function (e) {
+    if (directEditing.isActive()) {
+      directEditing.cancel();
+    }
+  });
 
-        if (directEditing.isActive()) {
-            directEditing.complete();
-        }
-    });
+  eventBus.on("directEditing.activate", function (event) {
+    resizeHandles.removeResizers();
+  });
 
-    // cancel on command stack changes
-    eventBus.on(['commandStack.changed'], function (e) {
-        if (directEditing.isActive()) {
-            directEditing.cancel();
-        }
-    });
+  eventBus.on("create.end", 500, function (event) {
+    var context = event.context,
+      element = context.shape,
+      canExecute = event.context.canExecute,
+      isTouch = event.isTouch;
 
+    // TODO(nikku): we need to find a way to support the
+    // direct editing on mobile devices; right now this will
+    // break for desworkflowediting on mobile devices
+    // as it breaks the user interaction workflow
 
-    eventBus.on('directEditing.activate', function (event) {
-        resizeHandles.removeResizers();
-    });
-
-    eventBus.on('create.end', 500, function (event) {
-
-        var context = event.context,
-            element = context.shape,
-            canExecute = event.context.canExecute,
-            isTouch = event.isTouch;
-
-        // TODO(nikku): we need to find a way to support the
-        // direct editing on mobile devices; right now this will
-        // break for desworkflowediting on mobile devices
-        // as it breaks the user interaction workflow
-
-        // TODO(nre): we should temporarily focus the edited element
-        // here and release the focused viewport after the direct edit
-        // operation is finished
-        if (isTouch) {
-            return;
-        }
-
-        if (!canExecute) {
-            return;
-        }
-
-        if (context.hints && context.hints.createElementsBehavior === false) {
-            return;
-        }
-
-        activateDirectEdit(element, false);
-    });
-
-    eventBus.on('autoPlace.end', 500, function (event) {
-        activateDirectEdit(event.shape, false);
-    });
-
-
-    function activateDirectEdit(element, force) {
-        if (force ||
-            isAny(element, ['rom:Role'])) {
-            directEditing.activate(element);
-        }
+    // TODO(nre): we should temporarily focus the edited element
+    // here and release the focused viewport after the direct edit
+    // operation is finished
+    if (isTouch) {
+      return;
     }
 
-}
+    if (!canExecute) {
+      return;
+    }
 
-LabelEditingProvider.$inject = [
-    'eventBus',
-    'odFactory',
-    'canvas',
-    'directEditing',
-    'modeling',
-    'resizeHandles',
-    'textRenderer'
-];
+    if (context.hints && context.hints.createElementsBehavior === false) {
+      return;
+    }
 
+    activateDirectEdit(element, false);
+  });
 
-/**
- * Activate direct editing for objects and text annotations.
- *
- * @param  {djs.model.Base} element
- *
- * @return {Object} an object with properties bounds (position and size), text and options
- */
-LabelEditingProvider.prototype.activate = function (element) {
+  eventBus.on("autoPlace.end", 500, function (event) {
+    activateDirectEdit(event.shape, false);
+  });
 
+  function activateDirectEdit(element, force) {
+    if (force || isAny(element, ["rom:Position", "rom:OrganizationalUnit"])) {
+    }
+  }
+
+  LabelEditingProvider.$inject = [
+    "eventBus",
+    "odFactory",
+    "canvas",
+    "directEditing",
+    "modeling",
+    "resizeHandles",
+    "textRenderer",
+  ];
+
+  /**
+   * Activate direct editing for objects and text annotations.
+   *
+   * @param  {djs.model.Base} element
+   *
+   * @return {Object} an object with properties bounds (position and size), text and options
+   */
+  LabelEditingProvider.prototype.activate = function (element) {
     // text
     var text = getLabel(element);
 
     if (text === undefined) {
-        return;
+      return;
     }
 
     var context = {
-        text: text
+      text: text,
     };
 
     // bounds
@@ -132,33 +134,32 @@ LabelEditingProvider.prototype.activate = function (element) {
     assign(context, bounds);
 
     var options = {
-        centerVertically: true
+      centerVertically: true,
     };
 
     // external labels
     if (isLabelExternal(element)) {
-        assign(options, {
-            autoResize: true
-        });
+      assign(options, {
+        autoResize: true,
+      });
     }
 
     assign(context, {
-        options: options
+      options: options,
     });
 
     return context;
-};
+  };
 
-
-/**
- * Get the editing bounding box based on the element's size and position
- *
- * @param  {djs.model.Base} element
- *
- * @return {Object} an object containing information about position
- *                  and size (fixed or minimum and/or maximum)
- */
-LabelEditingProvider.prototype.getEditingBBox = function (element) {
+  /**
+   * Get the editing bounding box based on the element's size and position
+   *
+   * @param  {djs.model.Base} element
+   *
+   * @return {Object} an object containing information about position
+   *                  and size (fixed or minimum and/or maximum)
+   */
+  LabelEditingProvider.prototype.getEditingBBox = function (element) {
     var canvas = this._canvas;
 
     var target = element.label || element;
@@ -166,117 +167,114 @@ LabelEditingProvider.prototype.getEditingBBox = function (element) {
     var bbox = canvas.getAbsoluteBBox(target);
 
     var mid = {
-        x: bbox.x + bbox.width / 2,
-        y: bbox.y + bbox.height / 2
+      x: bbox.x + bbox.width / 2,
+      y: bbox.y + bbox.height / 2,
     };
 
     // default position
-    var bounds = {x: bbox.x, y: bbox.y};
+    var bounds = { x: bbox.x, y: bbox.y };
 
     var zoom = canvas.zoom();
 
     var defaultStyle = this._textRenderer.getDefaultStyle(),
-        externalStyle = this._textRenderer.getExternalStyle();
+      externalStyle = this._textRenderer.getExternalStyle();
 
     // take zoom into account
     var externalFontSize = externalStyle.fontSize * zoom,
-        externalLineHeight = externalStyle.lineHeight,
-        defaultFontSize = defaultStyle.fontSize * zoom,
-        defaultLineHeight = defaultStyle.lineHeight;
+      externalLineHeight = externalStyle.lineHeight,
+      defaultFontSize = defaultStyle.fontSize * zoom,
+      defaultLineHeight = defaultStyle.lineHeight;
 
     var style = {
-        fontFamily: this._textRenderer.getDefaultStyle().fontFamily,
-        fontWeight: this._textRenderer.getDefaultStyle().fontWeight
+      fontFamily: this._textRenderer.getDefaultStyle().fontFamily,
+      fontWeight: this._textRenderer.getDefaultStyle().fontWeight,
     };
 
+    if (isAny(element, ["rom:Position", "rom:OrganizationalUnit"])) {
+      assign(bounds, {
+        width: bbox.width,
+        y: bbox.y,
+        height: bbox.height,
+      });
 
-    if (isAny(element, ['rom:Role'])) {
-
-        assign(bounds, {
-            width: bbox.width,
-            y: bbox.y,
-            height: bbox.height
-        });
-
-        assign(style, {
-            fontSize: defaultFontSize + 'px',
-            lineHeight: defaultLineHeight,
-            paddingTop: (7 * zoom) + 'px',
-            paddingBottom: (7 * zoom) + 'px',
-            paddingLeft: (5 * zoom) + 'px',
-            paddingRight: (5 * zoom) + 'px',
-            background: 'white'
-        });
+      assign(style, {
+        fontSize: defaultFontSize + "px",
+        lineHeight: defaultLineHeight,
+        paddingTop: 7 * zoom + "px",
+        paddingBottom: 7 * zoom + "px",
+        paddingLeft: 5 * zoom + "px",
+        paddingRight: 5 * zoom + "px",
+        background: "white",
+      });
     }
 
     var width = 90 * zoom,
-        paddingTop = 7 * zoom,
-        paddingBottom = 4 * zoom;
+      paddingTop = 7 * zoom,
+      paddingBottom = 4 * zoom;
 
     // external labels for events, data elements, gateways, groups and connections
     if (target.labelTarget) {
-        assign(bounds, {
-            width: width,
-            height: bbox.height + paddingTop + paddingBottom,
-            x: mid.x - width / 2,
-            y: bbox.y - paddingTop
-        });
+      assign(bounds, {
+        width: width,
+        height: bbox.height + paddingTop + paddingBottom,
+        x: mid.x - width / 2,
+        y: bbox.y - paddingTop,
+      });
 
-        assign(style, {
-            fontSize: externalFontSize + 'px',
-            lineHeight: externalLineHeight,
-            paddingTop: paddingTop + 'px',
-            paddingBottom: paddingBottom + 'px'
-        });
+      assign(style, {
+        fontSize: externalFontSize + "px",
+        lineHeight: externalLineHeight,
+        paddingTop: paddingTop + "px",
+        paddingBottom: paddingBottom + "px",
+      });
     }
 
     // external label not yet created
-    if (isLabelExternal(target)
-        && !hasExternalLabel(target)
-        && !isLabel(target)) {
+    if (
+      isLabelExternal(target) &&
+      !hasExternalLabel(target) &&
+      !isLabel(target)
+    ) {
+      var externalLabelMid = getExternalLabelMid(element);
 
-        var externalLabelMid = getExternalLabelMid(element);
+      var absoluteBBox = canvas.getAbsoluteBBox({
+        x: externalLabelMid.x,
+        y: externalLabelMid.y,
+        width: 0,
+        height: 0,
+      });
 
-        var absoluteBBox = canvas.getAbsoluteBBox({
-            x: externalLabelMid.x,
-            y: externalLabelMid.y,
-            width: 0,
-            height: 0
-        });
+      var height = externalFontSize + paddingTop + paddingBottom;
 
-        var height = externalFontSize + paddingTop + paddingBottom;
+      assign(bounds, {
+        width: width,
+        height: height,
+        x: absoluteBBox.x - width / 2,
+        y: absoluteBBox.y - height / 2,
+      });
 
-        assign(bounds, {
-            width: width,
-            height: height,
-            x: absoluteBBox.x - width / 2,
-            y: absoluteBBox.y - height / 2
-        });
-
-        assign(style, {
-            fontSize: externalFontSize + 'px',
-            lineHeight: externalLineHeight,
-            paddingTop: paddingTop + 'px',
-            paddingBottom: paddingBottom + 'px'
-        });
+      assign(style, {
+        fontSize: externalFontSize + "px",
+        lineHeight: externalLineHeight,
+        paddingTop: paddingTop + "px",
+        paddingBottom: paddingBottom + "px",
+      });
     }
 
-    return {bounds: bounds, style: style};
-};
+    return { bounds: bounds, style: style };
+  };
 
-
-LabelEditingProvider.prototype.update = function (element, newLabel) {
-
+  LabelEditingProvider.prototype.update = function (element, newLabel) {
     if (isEmptyText(newLabel)) {
-        newLabel = null;
+      newLabel = null;
     }
 
     this._modeling.updateLabel(element, newLabel);
-};
+  };
 
+  // helpers //////////////////////
 
-// helpers //////////////////////
-
-function isEmptyText(label) {
+  function isEmptyText(label) {
     return !label || !label.trim();
+  }
 }
