+++ app/lib/rolemodeler/features/ordering/ODOrderingProvider.js
@@ -1,11 +1,10 @@
-import inherits from 'inherits';
+import inherits from "inherits";
 
-import OrderingProvider from 'diagram-js/lib/features/ordering/OrderingProvider';
+import OrderingProvider from "diagram-js/lib/features/ordering/OrderingProvider";
 
-import {isAny} from '../../../common/features/modeling/ModelingUtil';
+import { isAny } from "../../../common/features/modeling/ModelingUtil";
 
-import {find, findIndex} from 'min-dash';
-
+import { find, findIndex } from "min-dash";
 
 /**
  * a simple ordering provider that makes sure:
@@ -14,119 +13,108 @@ import {find, findIndex} from 'min-dash';
  * (1) elements are ordered by a {level} property
  */
 export default function ODOrderingProvider(eventBus, canvas, translate) {
+  OrderingProvider.call(this, eventBus);
 
-    OrderingProvider.call(this, eventBus);
+  var orders = [
+    { type: "rom:Position", order: { level: 5 } },
+    { type: "rom:OrganizationalUnit", order: { level: 5 } },
+    { type: "rom:Link", order: { level: 3 } },
+  ];
 
-    var orders = [
-        {type: 'rom:Role', order: {level: 5}},
-        {
-            type: 'rom:Inheritance',
-            order: {
-                level: 3,
-            }
-        },
-    ];
-
-    function computeOrder(element) {
-        if (element.labelTarget) {
-            return {level: 10};
-        }
-
-        var entry = find(orders, function (o) {
-            return isAny(element, [o.type]);
-        });
-
-        return entry && entry.order || {level: 1};
+  function computeOrder(element) {
+    if (element.labelTarget) {
+      return { level: 10 };
     }
 
-    function getOrder(element) {
+    var entry = find(orders, function (o) {
+      return isAny(element, [o.type]);
+    });
 
-        var order = element.order;
+    return (entry && entry.order) || { level: 1 };
+  }
 
-        if (!order) {
-            element.order = order = computeOrder(element);
-        }
+  function getOrder(element) {
+    var order = element.order;
 
-        if (!order) {
-            throw new Error('no order for <' + element.id + '>');
-        }
-
-        return order;
+    if (!order) {
+      element.order = order = computeOrder(element);
     }
 
-    function findActualParent(element, newParent, containers) {
-
-        var actualParent = newParent;
-
-        while (actualParent) {
-
-            if (isAny(actualParent, containers)) {
-                break;
-            }
-
-            actualParent = actualParent.parent;
-        }
-
-        if (!actualParent) {
-            throw new Error(translate('no parent for {element} in {parent}', {
-                element: element.id,
-                parent: newParent.id
-            }));
-        }
-
-        return actualParent;
+    if (!order) {
+      throw new Error("no order for <" + element.id + ">");
     }
 
-    this.getOrdering = function (element, newParent) {
+    return order;
+  }
 
-        // render labels always on top
-        if (element.labelTarget) {
-            return {
-                parent: canvas.getRootElement(),
-                index: -1
-            };
-        }
+  function findActualParent(element, newParent, containers) {
+    var actualParent = newParent;
 
-        var elementOrder = getOrder(element);
+    while (actualParent) {
+      if (isAny(actualParent, containers)) {
+        break;
+      }
 
+      actualParent = actualParent.parent;
+    }
 
-        if (elementOrder.containers) {
-            newParent = findActualParent(element, newParent, elementOrder.containers);
-        }
+    if (!actualParent) {
+      throw new Error(
+        translate("no parent for {element} in {parent}", {
+          element: element.id,
+          parent: newParent.id,
+        })
+      );
+    }
 
+    return actualParent;
+  }
 
-        var currentIndex = newParent.children.indexOf(element);
+  this.getOrdering = function (element, newParent) {
+    // render labels always on top
+    if (element.labelTarget) {
+      return {
+        parent: canvas.getRootElement(),
+        index: -1,
+      };
+    }
 
-        var insertIndex = findIndex(newParent.children, function (child) {
+    var elementOrder = getOrder(element);
 
-            // do not compare with labels, they are created
-            // in the wrong order (right after elements) during import and
-            // mess up the positioning.
-            if (!element.labelTarget && child.labelTarget) {
-                return false;
-            }
+    if (elementOrder.containers) {
+      newParent = findActualParent(element, newParent, elementOrder.containers);
+    }
 
-            return elementOrder.level < getOrder(child).level;
-        });
+    var currentIndex = newParent.children.indexOf(element);
 
+    var insertIndex = findIndex(newParent.children, function (child) {
+      // do not compare with labels, they are created
+      // in the wrong order (right after elements) during import and
+      // mess up the positioning.
+      if (!element.labelTarget && child.labelTarget) {
+        return false;
+      }
 
-        // if the element is already in the child list at
-        // a smaller index, we need to adjust the insert index.
-        // this takes into account that the element is being removed
-        // before being re-inserted
-        if (insertIndex !== -1) {
-            if (currentIndex !== -1 && currentIndex < insertIndex) {
-                insertIndex -= 1;
-            }
-        }
+      return elementOrder.level < getOrder(child).level;
+    });
 
-        return {
-            index: insertIndex,
-            parent: newParent
-        };
+    // if the element is already in the child list at
+    // a smaller index, we need to adjust the insert index.
+    // this takes into account that the element is being removed
+    // before being re-inserted
+    if (insertIndex !== -1) {
+      if (currentIndex !== -1 && currentIndex < insertIndex) {
+        insertIndex -= 1;
+      }
+    }
+
+    return {
+      index: insertIndex,
+      parent: newParent,
     };
+  };
 }
 
-ODOrderingProvider.$inject = ['eventBus', 'canvas', 'translate'];
+ODOrderingProvider.$inject = ["eventBus", "canvas", "translate"];
 
 inherits(ODOrderingProvider, OrderingProvider);
