+++ app/lib/rolemodeler/roleLabelHandling/index.js
@@ -1,0 +1,6 @@
+import RoleLabelHandler from "./roleLabelHandler";
+
+export default {
+  __init__: ["roleLabelHandler"],
+  roleLabelHandler: ["type", RoleLabelHandler],
+};
